rename omni_connector_support
