+++ vllm_omni/engine/arg_utils.py
@@ -12,7 +12,7 @@ from vllm_omni.config import OmniModelConfig
 from vllm_omni.outputs.output_modality import OutputModality
 from vllm_omni.platforms import current_omni_platform
 from vllm_omni.plugins import load_omni_general_plugins
-from vllm_omni.worker.omni_connector_support import validate_worker_omni_connector
+from vllm_omni.worker.omni_connector_validation import validate_worker_omni_connector
 
 logger = init_logger(__name__)
